requirements.txt
python-dotenv==1.0.1
python-telegram-bot[all]==21.4
Requests==2.32.3
termcolor==2.4.0
